Reduce line-splitting
Unix makes roughly zero guarantees about when it will or won't
split or coalesce writes across pipes, which means that a single
write of a sentence of text could result in two separate calls
to the Write() method of an io.Writer used as Stdout or Stderr
for a command.

The least-bad solution, I think, is to have a separate goroutine
which waits for full lines, but will also time out after a bit
so that incomplete lines don't get lost forever, just delayed
very slightly (about a millisecond seems to cover everything).

Typical output change:

	09:13:02               web | npm
	09:13:02               web |  WARN
	09:13:02               web |  bootstrap@4.0.0 [...]

becomes:

	09:14:05               web | npm WARN bootstrap@4.0.0 [...]

This also addresses a subtle bug where single-byte non-newline
writes were getting discarded, but producing blank lines with
no header information.

As a secondary issue, instead of remaking the format string
for every message, we use a generic format string with a
length argument, since the format string never changes except
for that length.