Add mechanism for shutting down RPC server on shutdown
When the user cancels the program with ctrl+c or similar, we should
try to shut down the RPC server goroutine. Previously there was no
way to do that; now, we provide a context.Context that can be used to
cancel activity on the server.

The RPC server is given ten seconds for all of the in-flight
operations to complete, and then it's killed.

Eventually all of the operations should accept a context.Context as an
argument so they can similarly respond to external signals.