Add .travis.yml
The primary benefit of this is by building both Linux and Windows
machines, we test all of the code paths and ensure we're not checking
in incompatible code.

It would be good to add tests at some point but for the moment it's
a good idea to build the code for each GOOS before merging.