Add basic tests
These two tests check that we can start processes with goreman and
then send signals to them to terminate them.

Refactor startProc to accept the signal as an external argument, which
allows us to send incoming signals directly from the test.

Change the Travis CI configuration to just build the Windows config,
not actually test it, this fails now since we actually have tests that
can run.