Fix double-signal issue
Previously the server would take two Ctrl+C's to shut down - one to
trigger the logic in goreman.go (and shut down the RPC server), the
second to trigger the logic in startProcs. We should only handle
incoming signals in one place, instead of two - refactor the logic to
handle this case.

(I missed this during testing because we don't have tests around the
main function, only around the `start`/`startProcs` functions.)