Add support for running sub-processes under a PTY
This is really nice for running programs that output colour, such as
coloured log lines, etc.

I defaulted it to off because some applications when run under a PTY
will attempt to move the cursor around, clear lines, and so on, so it
should be up to the user to decide whether to enable it.